Extracted MenuRow component to own file
