+++ next-app/styles/Main.module.css
@@ -30,7 +30,7 @@
 }
 .main__mask--light {
     /*mask over bg to make more subtle*/
-    background: rgba(var(--white),0.7);
+    background: rgba(var(--white),0.6);
 }
 .main__mask--dark {
     /*mask over bg to make more subtle*/
